feat: create PasswordStrengthMeter component
