+++ frontend/src/app/components/elements/passwordStrengthMeter/PasswordStrengthMeter.module.scss
@@ -1,0 +1,45 @@
+.strengthLabel {
+  font-size: 14px;
+  margin-top: 5px;
+  color: #888;
+  text-align: left;
+  vertical-align: middle;
+
+  &__icon {
+    width: 1.2rem;
+    height: 1.2rem;
+    vertical-align: top;
+    &--x {
+      color: #FF3E3E;
+    }
+    &--check {
+      color: #4CAF50;
+    }
+  }
+
+  span {
+    font-weight: 500;
+  }
+}
+
+.strengthBarContainer {
+  width: 100%;
+  height: 8px;
+  background: #e0e0e0;
+  border-radius: 4px;
+  margin-top: 5px;
+  position: relative;
+  overflow: hidden;
+}
+
+.strengthBar {
+  height: 100%;
+  transition: width 0.3s ease-in-out, background-color 0.3s ease-in-out;
+}
+
+.passwordFeedback {
+  font-size: 12px;
+  color: #888;
+  margin-top: 4px;
+  margin-bottom: 1rem;
+}
